orders,offers changed

+++ api/place_order.php
@@ -46,7 +46,7 @@ if (empty($_POST['is_paid'])) {
 $user_id = $db->escapeString($_POST['user_id']);
 $buy_method = $db->escapeString($_POST['buy_method']);
 $is_paid = $db->escapeString($_POST['is_paid']);
-$sql = "SELECT product_id,quantity FROM cart WHERE user_id = '" . $user_id . "'";
+$sql = "SELECT cart.product_id AS product_id,cart.quantity AS quantity,products.discounted_price AS discounted_price,products.seller_id AS seller_id FROM cart,products WHERE cart.product_id = products.id AND user_id = $user_id";
 $db->sql($sql);
 $res = $db->getResult();
 $sql = "SELECT COUNT(id) AS count FROM cart WHERE user_id = '" . $user_id . "'";
@@ -58,22 +58,39 @@ if($is_paid == 'true'){
 }else{
     $payment_status = "UnPaid";
 }
+$sql = "SELECT * FROM delivery WHERE title = 'delivery'";
+$db->sql($sql);
+$resdeliver = $db->getResult();
 
-for ($i = 0; $i < $productcount; $i++) {
-    $product_id = $res[$i]['product_id'];
-    $quantity = $res[$i]['quantity'];
-    $discounted_price = $res[$i]['discounted_price'];
-    $delivery_charges = $res[$i]['delivery_charges'];
+if($productcount > 0){
+    for ($i = 0; $i < $productcount; $i++) {
+        $product_id = $res[$i]['product_id'];
+        $quantity = $res[$i]['quantity'];
+        $total_price = $res[$i]['quantity'] * $res[$i]['discounted_price'];
+        
+        $delivery_charges = $resdeliver[0]['charges'];
+        $seller_id = $res[$i]['seller_id'];
+        
+        $sql = "INSERT INTO orders(`user_id`,`product_id`,`seller_id`,`quantity`,`status`,`delivery_charges`,`buy_method`,`payment_status`,`total`)VALUES('$user_id','$product_id','$seller_id','$quantity','Received',$delivery_charges,'$buy_method','$payment_status','$total_price')";
+        $db->sql($sql);
+        $sql = "DELETE FROM cart WHERE product_id = '" . $product_id . "' AND user_id = '$user_id'";
+        $db->sql($sql);
     
-    $sql = "INSERT INTO orders(`user_id`,`product_id`,`quantity`,`status`,`delivery_charges`,`buy_method`,`payment_status`,`total`)VALUES('$user_id','$product_id','$quantity','Received',$delivery_charges,'$buy_method','$payment_status','$discounted_price')";
-    $db->sql($sql);
-    $sql = "DELETE FROM cart WHERE product_id = '" . $product_id . "' AND user_id = '$user_id'";
-    $db->sql($sql);
+    }
+    $response['success'] = true;
+    $response['message'] = "Order Placed Successfully";
+
+}
+else {
+    $response['success'] = false;
+    $response['message'] = "No Products Found In Cart";
 
 }
 
-$response['success'] = true;
-$response['message'] = "Order Placed Successfully";
+
+
+
+
 print_r(json_encode($response));
